Ensure a shallow compare remains shallow
Consider the case where either the mode and the size of a file is
identical but the modification time isn't. This means that the all
guard-if-clauses would be skipped and an actual deep compare would be
performed.

To ensure that a shallow copy indeed is shallow, perform:
- a check if both files are regular files;
- if the compare is shallow perform a diff of the file signatures (which
  include the mode, size and the modification time) and return true if
  the signatures match and false if they don't.

This second evaluation ensures that a shallow copy won't result into a
deep compare.