Simulate filecmp corner-case for shallow compare
A shallow compare takes the mode, size and mtime of a file to perform a
compare between both files. However, in case the mtime is the only
difference between 2 files, the filecmp method performs a deep compare
under the hood, which is in conflict with the `shallow` argument.

The new test verifies that when only the mtime differs no deep compare
is performed. If a deep compare would be performed (i.e. bit-by-bit) the
files would be considered equal, but if only a shallow copy is
performed, the compare should result in false. This new test verifies
this behaviour.